+++ server/api/handlers/group_update.go
@@ -59,7 +59,7 @@ func (h *Handler) UpdateGroup(c echo.Context) error {
 	if !ok {
 		return &echo.HTTPError{Code: http.StatusNotAcceptable, Message: "wrong token or missing info in token claims"}
 	}
-	if err := h.DB.Model(&models.User{}).Where("id = ?", uid).First(&u).Error; err != nil {
+	if err := h.DB.Model(&models.User{}).Where("id = ?", uint(uid)).First(&u).Error; err != nil {
 		return &echo.HTTPError{Code: http.StatusForbidden, Message: "wrong user attempting to update group"}
 	}
